Enables arXiv link processing and UI
Adds functionality to process arXiv links, scraping paper metadata and queueing the download of the PDF.
Improves the web interface to display a dashboard with the status of uploaded/linked papers.
The API has been modified to accept URL requests which will scrape the arXiv page for the PDF.
The API now also adds the id to the JobStatusResponse.

Adds react-query-devtools for easier debugging of web requests.